eodhd_news_replication/requirements.txt
# EODHD News Event Study Requirements
# Python 3.8+

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# API and Web
requests>=2.28.0
python-dotenv>=0.20.0

# Text Processing
nltk>=3.8.0
scikit-learn>=1.2.0

# Topic Modeling and ML
gensim>=4.2.0  # Alternative to sklearn LDA if needed

# Time Series and Statistics
statsmodels>=0.13.5

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0  # Optional for interactive plots

# Configuration
PyYAML>=6.0

# Progress bars (optional but helpful)
tqdm>=4.65.0

# Data storage
pyarrow>=10.0.0  # For parquet support
tables>=3.7.0  # Optional for HDF5 support

# Jupyter support (optional for notebooks)
jupyter>=1.0.0
ipykernel>=6.19.0
notebook>=6.5.0

# Testing (optional)
pytest>=7.2.0
pytest-cov>=4.0.0

# Code quality (optional)
black>=22.12.0
flake8>=6.0.0
mypy>=0.991